Resolve #8
